requirements.txt
# Discord bot
discord.py>=2.3.0

# Database
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.0
asyncpg>=0.29.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Utilities
python-dotenv>=1.0.0
